Moved tests for better project Structure
