+++ pom.xml
@@ -43,7 +43,7 @@
     </dependencies>
     <build>
         <!-- Plugin for running JUnit 5 tests with Maven -->
-        <testSourceDirectory>${project.basedir}/src/main/java/test/</testSourceDirectory>
+        <testSourceDirectory>src/main/test/</testSourceDirectory>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
