+++ src/main/test/db/DatabaseTestHelper.java
@@ -1,4 +1,4 @@
-package test.db;
+package db;
 
 import java.sql.Connection;
 import java.sql.Date;
